feat(core): Phase 5.10 - consolidate ZeddaError imports across modules
Consolidate ZeddaError references in _clean.py and _engine.py to point directly to canonical _errors.py. Single canonical definition maintained in _errors.py with backward compatibility preserved via __init__.py.
